Move ThrottleAdminEmailHandler to logging module
